+++ CHANGELOG.md
@@ -1,5 +1,4 @@
 # Estrogen `@VERSION@` `1.21.1` `Fabric & NeoForge`
 
 ### Changelog:
-- Temporary Boobs persisting beyond the grave
-- some Damage Type Tags missing for damage types
+- Sable Compatibility with Dream Blocks